app/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "portfolio"
version = "0.0.1"
authors = [
    {name = "Joseph Monyoro", email = "monyorojoseph@gmail.com"},
]
description = "my portfolio app"
requires-python = ">=3.8"
dependencies = [
    "Django==5.0.3",
    "python-dotenv==1.0.1",
    "gunicorn==21.2.0",
    "django-ninja==1.1.0",
    "django-ninja-extra==0.20.4",
    "django-ninja-jwt==5.3.0",
    "pillow==10.2.0",
    "psycopg2-binary==2.9.9",
    "django-cors-headers==4.3.1",
    "dj-database-url==2.1.0"
]

[tool.setuptools]
packages = ["setup", "core"]
